requirements.txt
telethon==1.34.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
